feat: add useFetcher hook to simplify Fetcher class creation
