+++ packages/frontend/src/hooks/useLatestKPITokens.ts
@@ -1,11 +1,7 @@
-import { Fetcher } from "@carrot-kpi/sdk";
-import { useProvider, useNetwork } from "wagmi";
+import { useNetwork } from "wagmi";
 import { KPIToken } from "@carrot-kpi/sdk";
-import {
-    usePreferDecentralization,
-    useIPFSGatewayURL,
-} from "@carrot-kpi/react";
 import { useQuery } from "@tanstack/react-query";
+import { useFetcher } from "@carrot-kpi/react";
 
 export const LATEST_KPI_TOKEN_QUERY_KEY_PREFIX = "latestKPITokens" as string;
 
@@ -13,15 +9,8 @@ export function useLatestKPITokens(limit = 5): {
     loading: boolean;
     kpiTokens: KPIToken[];
 } {
-    const preferDecentralization = usePreferDecentralization();
-    const ipfsGatewayURL = useIPFSGatewayURL();
     const { chain } = useNetwork();
-    const provider = useProvider();
-    const fetcher = new Fetcher({
-        provider,
-        ipfsGatewayURL,
-        preferDecentralization,
-    });
+    const { fetcher } = useFetcher();
 
     const { data: kpiTokens, isLoading: loading } = useQuery({
         queryKey: [LATEST_KPI_TOKEN_QUERY_KEY_PREFIX, { limit, chain }],
