+++ packages/react/src/hooks/useKPIToken.ts
@@ -1,27 +1,18 @@
 import { useEffect, useState } from "react";
-import { KPIToken, Fetcher } from "@carrot-kpi/sdk";
-import { useProvider, useNetwork } from "wagmi";
-import { usePreferDecentralization } from "./usePreferDecentralization";
-import { useIPFSGatewayURL } from "./useIPFSGatewayURL";
+import { KPIToken } from "@carrot-kpi/sdk";
+import { useNetwork } from "wagmi";
+import { useFetcher } from "./useFetcher";
 
 export function useKPIToken(kpiTokenAddress?: string): {
     loading: boolean;
     kpiToken: KPIToken | null;
 } {
-    const preferDecentralization = usePreferDecentralization();
-    const ipfsGatewayURL = useIPFSGatewayURL();
     const { chain } = useNetwork();
-    const provider = useProvider();
     const [kpiToken, setKPIToken] = useState<KPIToken | null>(null);
     const [loading, setLoading] = useState(true);
+    const { fetcher } = useFetcher();
 
     useEffect(() => {
-        const fetcher = new Fetcher({
-            provider,
-            ipfsGatewayURL,
-            preferDecentralization,
-        });
-
         let cancelled = false;
         async function fetchData(): Promise<void> {
             if (!chain || !kpiTokenAddress) return;
@@ -47,13 +38,7 @@ export function useKPIToken(kpiTokenAddress?: string): {
         return () => {
             cancelled = true;
         };
-    }, [
-        chain,
-        ipfsGatewayURL,
-        kpiTokenAddress,
-        preferDecentralization,
-        provider,
-    ]);
+    }, [chain, fetcher, kpiTokenAddress]);
 
     return { loading, kpiToken };
 }
