+++ packages/react/src/hooks/useKPITokenTemplates.ts
@@ -1,28 +1,19 @@
 import { useEffect, useState } from "react";
-import { Template, Fetcher } from "@carrot-kpi/sdk";
-import { useProvider, useNetwork } from "wagmi";
+import { Template } from "@carrot-kpi/sdk";
+import { useNetwork } from "wagmi";
 import { BigNumberish } from "@ethersproject/bignumber";
-import { usePreferDecentralization } from "./usePreferDecentralization";
-import { useIPFSGatewayURL } from "./useIPFSGatewayURL";
+import { useFetcher } from "./useFetcher";
 
 export function useKPITokenTemplates(ids?: BigNumberish[]): {
     loading: boolean;
     templates: Template[];
 } {
-    const preferDecentralization = usePreferDecentralization();
-    const ipfsGatewayURL = useIPFSGatewayURL();
     const { chain } = useNetwork();
-    const provider = useProvider();
     const [templates, setTemplates] = useState<Template[]>([]);
     const [loading, setLoading] = useState(true);
+    const { fetcher } = useFetcher();
 
     useEffect(() => {
-        const fetcher = new Fetcher({
-            provider,
-            ipfsGatewayURL,
-            preferDecentralization,
-        });
-
         let cancelled = false;
         async function fetchData(): Promise<void> {
             if (!chain) return;
@@ -42,7 +33,7 @@ export function useKPITokenTemplates(ids?: BigNumberish[]): {
         return () => {
             cancelled = true;
         };
-    }, [chain, ids, ipfsGatewayURL, preferDecentralization, provider]);
+    }, [chain, ids, fetcher]);
 
     return { loading, templates };
 }
